+++ doc/news.rst
@@ -30,6 +30,8 @@ you should review this document and see if your consfig needs updating.
   lists.  On Debian bullseye and older this will cause apt updates to emit
   harmless warnings.
 
+- Add LXC:USER-CONTAINER-STOPPED.
+
 1.2.0 (2022-11-29)
 ------------------